A bit of refactoring/cleanup.
This makes the code for the mixc and mixs commands consistent in
style and form.
